Resized all images, game window and updated code to get sprites 45*45 (and accomodate characters images)
Created character class and implemented movement (not limited yet)

Redefined the level generation within the class definition.

To review: first level generation returns initial position of character,
this is not used in display function. see if we can display originally
using the associated function (not repeating code) and extract the char position